change snapshot size to bigint and delete images when deleting test bubble

+++ workloadmgr/db/sqlalchemy/migrate_repo/versions/001_workloadmgr_init.py
@@ -4,7 +4,7 @@
 # All Rights Reserved.
 
 from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Text
-from sqlalchemy import Integer, MetaData, String, Table, UniqueConstraint
+from sqlalchemy import Integer, BigInteger, MetaData, String, Table, UniqueConstraint
 import sqlalchemy
 
 import pickle
@@ -145,8 +145,8 @@ def upgrade(migrate_engine):
         Column('snapshot_type', String(length=32), primary_key=False, nullable= False),
         Column('display_name', String(length=255)),
         Column('display_description', String(length=255)),
-        Column('size', Integer, nullable=False),
-        Column('uploaded_size', Integer, nullable=False),  
+        Column('size', BigInteger, nullable=False),
+        Column('uploaded_size', BigInteger, nullable=False),  
         Column('progress_percent', Integer, nullable=False),   
         Column('progress_msg', String(length=255)),
         Column('error_msg', String(length=255)),
@@ -164,7 +164,7 @@ def upgrade(migrate_engine):
         Column('vm_id', String(length=255), nullable= False),
         Column('vm_name', String(length=255), nullable= False),
         Column('snapshot_id', String(length=255), ForeignKey('snapshots.id')),
-        Column('size', Integer, nullable=False),
+        Column('size', BigInteger, nullable=False),
         Column('status', String(length=32), nullable=False),
         mysql_engine='InnoDB'
     )
@@ -192,7 +192,7 @@ def upgrade(migrate_engine):
         Column('resource_type', String(length=255)),
         Column('resource_name', String(length=255)),
         Column('resource_pit_id', String(length=255)),
-        Column('size', Integer, nullable=False),                
+        Column('size', BigInteger, nullable=False),                
         Column('status', String(length=32), nullable=False),
         UniqueConstraint('vm_id', 'snapshot_id', 'resource_name'),
         mysql_engine='InnoDB'
@@ -224,7 +224,7 @@ def upgrade(migrate_engine):
         Column('top', Boolean, default=False),
         Column('vault_service_url', String(4096)),    
         Column('vault_service_metadata', String(4096)),
-        Column('size', Integer, nullable=False),         
+        Column('size', BigInteger, nullable=False),         
         Column('status', String(length=32), nullable=False),
         mysql_engine='InnoDB'
     )        
@@ -282,8 +282,8 @@ def upgrade(migrate_engine):
         Column('restore_type', String(length=32), primary_key=False, nullable= False),
         Column('display_name', String(length=255)),
         Column('display_description', String(length=255)),
-        Column('size', Integer, nullable=False),             
-        Column('uploaded_size', Integer, nullable=False),          
+        Column('size', BigInteger, nullable=False),             
+        Column('uploaded_size', BigInteger, nullable=False),          
         Column('progress_percent', Integer, nullable=False),   
         Column('progress_msg', String(length=255)),
         Column('error_msg', String(length=255)),
